Numerous improvements:
- Fixed bug that always resulted in pLast (echo metric) to be 0
- Fixed casting warnings related to np.nan and integer types
- added test running echo metrics

+++ src/pyForMetrix/utils/rasterizer.py
@@ -62,7 +62,7 @@ class Rasterizer:
 
         # Prepare list for every output voxel
         XVoxelContains = [[] for i in range(noVoxel)]
-        XClosestIndex = np.full((noVoxel,), np.nan, dtype=int)
+        XClosestIndex = np.full((noVoxel,), np.nan, dtype=float)
 
         # Select points nearest to voxel centers - --------------------------------------
